Fix clang build
BUG=none
TEST=none
TBR=gman@chromium.org

Review URL: http://codereview.chromium.org/7493006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@93547 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f7e61316fae89776c8406073b17167fa3e991114

+++ gpu/command_buffer/client/gles2_implementation_unittest.cc
@@ -1370,7 +1370,9 @@ static bool CheckRect(
     return false;
   }
 
-  int r2_stride = flip_y ? -static_cast<int>(padded_row_size) : padded_row_size;
+  int r2_stride = flip_y ?
+      -static_cast<int>(padded_row_size) :
+      static_cast<int>(padded_row_size);
   r2 = flip_y ? (r2 + (height - 1) * padded_row_size) : r2;
 
   for (int y = 0; y < height; ++y) {
